Shorten pathnames of generator tests.
