+++ src/System.Windows.Forms.Analyzers/cs/tests/UnitTests/Generators/ApplicationConfiguration/ProjectFileReaderTests.FontConverter.cs
