+++ src/System.Windows.Forms.Analyzers/cs/tests/UnitTests/Generators/ApplicationConfiguration/TestData/GenerateInitialize_default_boilerplate.cs
