+++ src/System.Windows.Forms.Analyzers/cs/tests/UnitTests/Generators/ApplicationConfiguration/TestData/GenerateInitialize_user_top_level.cs
