+++ src/System.Windows.Forms.Analyzers/cs/tests/UnitTests/Generators/ApplicationConfiguration/Verifiers/CSharpSourceGeneratorTest`1.cs
