+++ src/System.Windows.Forms.Analyzers/cs/tests/UnitTests/Generators/ApplicationConfiguration/Verifiers/CSharpVerifierHelper.cs
