+++ src/System.Windows.Forms.Analyzers/cs/tests/UnitTests/System.Windows.Forms.Analyzers.CSharp.Tests.csproj
@@ -29,8 +29,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Remove="Generators\ApplicationConfigurationGeneratorTests\MockData\*.cs" />
-    <None Include="Generators\ApplicationConfigurationGeneratorTests\MockData\*.*">
+    <Compile Remove="Generators\ApplicationConfiguration\TestData\*.cs" />
+    <None Include="Generators\ApplicationConfiguration\TestData\*.*">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
