+++ src/System.Windows.Forms.Analyzers/generic/tests/UnitTests/Analyzers/AppManifestAnalyzerTests.cs
