+++ src/System.Windows.Forms.Analyzers/generic/tests/UnitTests/Analyzers/ApplicationConfigTests.GraphicsUnit.cs
