+++ src/System.Windows.Forms.Analyzers/generic/tests/UnitTests/Analyzers/TestData/invalid.manifest
