+++ src/System.Windows.Forms.Analyzers/generic/tests/UnitTests/Analyzers/TestData/nodpi.manifest
