+++ src/System.Windows.Forms.Analyzers/generic/tests/UnitTests/Microsoft.WinForms/TestFileEntry.cs
