+++ src/System.Windows.Forms.Analyzers/generic/tests/UnitTests/Microsoft.WinForms/TestFileLoader.cs
